Make the craft rate fractional so an 80-bar item can be served at all
ForgeTicks puts an Elder rune platebody + 5 at 10136 ticks for its 80
bars. ActionsPerHour then computed 1*6000/10157 by integer division and
returned 0, evaluate's 1/aph became +Inf, json.Marshal refused the
document, and gin recorded the render error and aborted - so
GET /calc/45742?player=Zezima answered HTTP 200 with a zero-byte body
and logged it as a success. On /calc/batch one such id emptied the whole
response, inverting the per-entry error contract that route is built on.

The quantity is genuinely fractional - that item takes about 1.7 hours -
so ActionsPerHour returns float64, chooseAPH carries the float through,
evaluate derives hours from it, and actions_per_hour (and a step's aph)
is a JSON number. Flooring the rate at 1 would have overstated the item
by seventy percent, which is the class of invented number this work
exists to remove. Breaking change to a field older than this branch,
acceptable while it is unmerged and nothing consumes the API.

Defence in depth: a path whose figures are not finite is now dropped as
unpriceable, whatever produced them, so no infinity can reach the
serialiser again through some later route. Its test does not depend on
the arithmetic above - it drives the rate to zero through an
unconfigured house default instead.

The trip count stays as it is: charging one bank trip for a craft that
needs three inventories understates the largest craft in the game by 21
ticks in 10157, and correcting it changes behaviour for every oversized
craft rather than closing this defect. Recorded as follow-up.

+++ calc-service/internal/service/top.go
@@ -63,7 +63,7 @@ type TopRow struct {
 	// otherwise.
 	GPPerHourLimited float64 `json:"gp_per_hour_limited"`
 	GPPerXP          float64 `json:"gp_per_xp"`
-	ActionsPerHour   int     `json:"actions_per_hour"`
+	ActionsPerHour   float64 `json:"actions_per_hour"`
 	APHSource        string  `json:"aph_source"`
 	Complete         bool    `json:"complete"`
 	BindingItemName  string  `json:"binding_item_name,omitempty"`
